# public/swagger/openapi.yml
servers:
- url: http://localhost:5000
  description: Development server
- url: http://localhost:5000
  description: Development server
paths:
...
  /v1/product:
    get:
      summary: Product List
      parameters:
      - name: limit
        in: query
        description: List limit
        required: false
        schema:
          type: int
          example: 20
      - name: offset
        in: query
        description: List offset
        required: false
        schema:
          type: int
          example: 0
      - name: fields
        in: query
        description: Filter fields with comma
        required: false
        schema:
          type: string
          example: null
      - name: order_by
        in: query
        description: Ordination of list
        required: false
        schema:
          type: string
          enum:
          - asc
          - desc
      - name: sort_by
        in: query
        description: Sorting of the list
        required: false
        schema:
          type: string
          example: id
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HateosProductListResponse'
        4xx:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListErrorResponse'
        5xx:
          description: Service fail response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListErrorResponse'
    post:
      summary: Product Create
      requestBody:
        description: Product to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCreateResponse'
        4xx:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCreateErrorResponse'
        5xx:
          description: Service fail response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCreateErrorResponse'
  /v1/product/{uuid}:
    get:
      summary: Product Get
      parameters:
      - in: path
        name: uuid
        description: Product Id
        required: true
        schema:
          type: string
          format: uuid
          example: 4bcad46b-6978-488f-8153-1c49f8a45244
      - name: fields
        in: query
        description: Filter fields with comma
        required: false
        schema:
          type: string
          example: null
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HateosProductGetResponse'
        4xx:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductGetErrorResponse'
        5xx:
          description: Service fail response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductGetErrorResponse'
    put:
      summary: Complete Product Update
      parameters:
      - in: path
        name: uuid
        description: Product Id
        required: true
        schema:
          type: string
          format: uuid
          example: 4bcad46b-6978-488f-8153-1c49f8a45244
      requestBody:
        description: Product to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCompleteUpdateRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductUpdateResponse'
        4xx:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductUpdateErrorResponse'
        5xx:
          description: Service fail response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductUpdateErrorResponse'
    patch:
      summary: Soft Product Update
      parameters:
      - in: path
        name: uuid
        description: Product Id
        required: true
        schema:
          type: string
          format: uuid
          example: 4bcad46b-6978-488f-8153-1c49f8a45244
      requestBody:
        description: Product field to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductSoftUpdateRequest'
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductUpdateResponse'
        4xx:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductUpdateErrorResponse'
        5xx:
          description: Service fail response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductUpdateErrorResponse'
    delete:
      summary: Soft Product Delete
      parameters:
      - in: path
        name: uuid
        description: Product Id
        required: true
        schema:
          type: string
          format: uuid
          example: 4bcad46b-6978-488f-8153-1c49f8a45244
      responses:
        '200':
          description: Success response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSoftDeleteResponse'
        4xx:
          description: Error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSoftDeleteErrorResponse'
        5xx:
          description: Service fail response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductSoftDeleteErrorResponse'
info:
  title: serverless-python-template
  version: 1.0.0
openapi: 3.0.2
components:
  schemas:
    RequestControl:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        count:
          type: integer
        limit:
          type: integer
      required:
      - limit
    Meta:
      type: object
      properties:
        href:
          type: string
          format: url
        next:
          type: string
          format: url
        previous:
          type: string
          format: url
        last:
          type: string
          format: url
        first:
          type: string
          format: url
...
    HateosProductListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        control:
          $ref: '#/components/schemas/RequestControl'
        message:
          type: string
          example: Success
        params:
          type: array
          items:
            type: string
        label:
          type: string
          example: common.success
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        code:
          type: integer
          example: 1
      required:
      - code
    ProductListErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        trace:
          type: string
        message:
          type: string
          example: Unable to return the list data, please review your request
        params:
          type: array
          items:
            type: string
        label:
          type: string
          example: common.error.list_error
        details:
          type: string
        code:
          type: integer
          example: 11
      required:
      - code
    Link:
      type: object
      properties:
        href:
          type: string
        method:
          type: string
        rel:
          type: string
    HateosProductGetResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Success
        params:
          type: array
          items:
            type: string
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
        label:
          type: string
          example: common.success
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          $ref: '#/components/schemas/Product'
        code:
          type: integer
          example: 1
      required:
      - code
    ProductGetErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        trace:
          type: string
        message:
          type: string
          example: Unable to find the record
        params:
          type: array
          items:
            type: string
        label:
          type: string
          example: common.error.find_error
        details:
          type: string
        code:
          type: integer
          example: 14
      required:
      - code
    ProductCreateRequest:
      type: object
      properties:
        sku:
          type: integer
          example: 1
        active:
          type: integer
          enum:
          - 0
          - 1
        description:
          type: string
          example: Common Pencil description
        name:
          type: string
          example: Common Pencil
        supplier_id:
          type: integer
          example: 1
    ProductCreateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Success
        params:
          type: array
          items:
            type: string
        label:
          type: string
          example: common.success
        data:
          $ref: '#/components/schemas/Product'
        code:
          type: integer
          example: 1
      required:
      - code
    ProductCreateErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        trace:
          type: string
        message:
          type: string
          example: Unable to create the record
        params:
          type: array
          items:
            type: string
        label:
          type: string
          example: common.error.create_error
        details:
          type: string
        code:
          type: integer
          example: 17
      required:
      - code
    ProductCompleteUpdateRequest:
      type: object
      properties:
        sku:
          type: integer
          example: 1
        active:
          type: integer
          enum:
          - 0
          - 1
        description:
          type: string
          example: Common Pencil description
        name:
          type: string
          example: Common Pencil
        supplier_id:
          type: integer
          example: 1
    ProductUpdateResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Success
        params:
          type: array
          items:
            type: string
        label:
          type: string
          example: common.success
        data:
          $ref: '#/components/schemas/Product'
        code:
          type: integer
          example: 1
      required:
      - code
    ProductUpdateErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        trace:
          type: string
        message:
          type: string
          example: Unable to update the record
        params:
          type: array
          items:
            type: string
        label:
          type: string
          example: common.error.update_error
        details:
          type: string
        code:
          type: integer
          example: 18
      required:
      - code
    ProductSoftUpdateRequest:
      type: object
      properties:
        field:
          type: string
          example: value
    ProductSoftDeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Success
        params:
          type: array
          items:
            type: string
        label:
          type: string
          example: common.success
        data:
          type: object
          example:
            deleted: true
        code:
          type: integer
          example: 1
      required:
      - code
    ProductSoftDeleteErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        trace:
          type: string
        message:
          type: string
          example: Unable to update the record
        params:
          type: array
          items:
            type: string
        label:
          type: string
          example: common.error.update_error
        details:
          type: string
        code:
          type: integer
          example: 18
      required:
      - code